Update the rss reader, css and index html
For the rss reader, a small introduction has been added which explains the use of the rss reader itself. In the css, the styling has been made for this. Now for the index html, the logo button at the top has been made clickable which is connected to the homepage.